Reorganized Root/Routes into components folder
